requirements.txt
pyspark==3.5.0
psutil==5.9.6
boto3==1.33.10
pandas==2.2.0
azure-storage-blob==12.19.0
python-dotenv==1.0.1